feat: implemented docker

+++ src/main/java/com/example/demo/service/TaskService.java
@@ -3,10 +3,13 @@ package com.example.demo.service;
 import java.util.List;
 import java.util.Optional;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import com.example.demo.DTO.request.TaskRequestDTO;
 import com.example.demo.DTO.request.TaskUpdateRequestDTO;
+import com.example.demo.exception.TaskInternalServerErrorException;
+import com.example.demo.exception.TaskMissingRequestBodyException;
 import com.example.demo.exception.TaskNotFoundException;
 import com.example.demo.model.Task;
 import com.example.demo.repository.TaskRepository;
@@ -19,40 +22,84 @@ public class TaskService {
     private final TaskRepository taskRepository;
 
     public List<Task> getAllTasks() {
-        List<Task> tasks = taskRepository.findAll();
+        try {
+            List<Task> tasks = taskRepository.findAll();
 
-        return tasks;
+            return tasks;
+        } catch (Exception ex) {
+            throw new TaskInternalServerErrorException("Internal server error", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     public Optional<Task> getById(int id) {
-        Optional<Task> task = taskRepository.findById(id);
+        try {
+            Optional<Task> task = taskRepository.findById(id);
 
-        return task;
+            if (task.isEmpty()) {
+                throw new TaskNotFoundException("Task not found with this id", HttpStatus.NOT_FOUND);
+            }
+
+            return task;
+        } catch (TaskNotFoundException ex) {
+            throw ex;
+        } catch (Exception ex) {
+            throw new TaskInternalServerErrorException("Internal server error", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+
     }
 
     public Task createTask(TaskRequestDTO taskRequestDTO) {
-        Task taskToCreate = new Task();
-        taskToCreate.setTitle(taskRequestDTO.getTitle());
-        taskToCreate.setCompleted(taskRequestDTO.getCompleted());
+        try {
+            Task taskToCreate = new Task();
 
-        return taskRepository.save(taskToCreate);
+            if (taskRequestDTO.getTitle().isEmpty()) {
+                throw new TaskMissingRequestBodyException("The filed title is required", HttpStatus.BAD_REQUEST);
+            }
+
+            taskToCreate.setTitle(taskRequestDTO.getTitle());
+            taskToCreate.setCompleted(taskRequestDTO.getCompleted());
+
+            return taskRepository.save(taskToCreate);
+        } catch (TaskMissingRequestBodyException ex) {
+            throw ex;
+        } catch (Exception ex) {
+            throw new TaskInternalServerErrorException("Internal server error", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     public Task updateTask(TaskUpdateRequestDTO taskUpdateRequestDTO, int id) {
-        Optional<Task> taskToUpdate = taskRepository.findById(id);
+        try {
+            Optional<Task> taskToUpdate = taskRepository.findById(id);
         
-        if (taskToUpdate.isEmpty()) {
-            throw new TaskNotFoundException("Task not foud with this id");
+            if (taskToUpdate.isEmpty()) {
+                throw new TaskNotFoundException("Task not foud with this id", HttpStatus.NOT_FOUND);
+            }
+
+            taskToUpdate.get().setTitle(taskUpdateRequestDTO.getTitle().isEmpty() ? taskToUpdate.get().getTitle() : taskUpdateRequestDTO.getTitle());
+            taskToUpdate.get().setCompleted(taskUpdateRequestDTO.getCompleted() != null ? taskUpdateRequestDTO.getCompleted() : taskToUpdate.get().getCompleted());
+
+            return taskRepository.save(taskToUpdate.get());
+        } catch (TaskNotFoundException ex) {
+            throw ex;
+        } catch (Exception ex) {
+            throw new TaskInternalServerErrorException("Internal server error", HttpStatus.INTERNAL_SERVER_ERROR);
         }
-
-        taskToUpdate.get().setTitle(taskUpdateRequestDTO.getTitle().isEmpty() ? taskToUpdate.get().getTitle() : taskUpdateRequestDTO.getTitle());
-        taskToUpdate.get().setCompleted(taskUpdateRequestDTO.getCompleted() != null ? taskUpdateRequestDTO.getCompleted() : taskToUpdate.get().getCompleted());
-
-        return taskRepository.save(taskToUpdate.get());
+      
     }
 
     public void delteById(int id) {
-        Optional<Task> task = taskRepository.findById(id);
-        taskRepository.delete(task.get());
+        try {
+            Optional<Task> task = taskRepository.findById(id);
+
+            if (task.isEmpty()) {
+                throw new TaskNotFoundException("Task not foud with this id", HttpStatus.NOT_FOUND);
+            }
+
+            taskRepository.delete(task.get());
+        } catch (TaskNotFoundException ex) {
+            throw ex;
+        } catch (Exception ex) {
+            throw new TaskInternalServerErrorException("Internal server error", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 }
